[REFACTOR] extension 파일 분리
